test(questionnaire-web-app): move empty appeal to tests as only used there
