+++ packages/lpa-questionnaire-web-app/tests/unit/controllers/other-appeals.test.js
@@ -2,7 +2,7 @@ const otherAppealsController = require('../../../src/controllers/other-appeals')
 const { createOrUpdateAppealReply } = require('../../../src/lib/appeal-reply-api-wrapper');
 const { getTaskStatus } = require('../../../src/services/task.service');
 const logger = require('../../../src/lib/logger');
-const appealReply = require('../../../src/lib/empty-appeal-reply');
+const appealReply = require('../emptyAppealReply');
 const { VIEW } = require('../../../src/lib/views');
 const { mockReq, mockRes } = require('../mocks');
