+++ packages/lpa-questionnaire-web-app/tests/unit/controllers/upload-plans.test.js
@@ -8,7 +8,7 @@ const {
 const logger = require('../../../src/lib/logger');
 const { createOrUpdateAppealReply } = require('../../../src/lib/appeal-reply-api-wrapper');
 const { VIEW } = require('../../../src/lib/views');
-const emptyAppealReply = require('../../../src/lib/empty-appeal-reply');
+const emptyAppealReply = require('../emptyAppealReply');
 const { getTaskStatus } = require('../../../src/services/task.service');
 const { mockReq, mockRes } = require('../mocks');
