+++ packages/lpa-questionnaire-web-app/tests/unit/mocks.js
@@ -1,7 +1,7 @@
 jest.mock('../../src/lib/logger');
 
 const logger = require('../../src/lib/logger');
-const emptyAppealReply = require('../../src/lib/empty-appeal-reply');
+const emptyAppealReply = require('./emptyAppealReply');
 
 const mockReq = (appealReply = emptyAppealReply, id = 'mock-id') => ({
   log: logger,
